tutorial/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pillow>=8.0.0
matplotlib>=3.5.0
imageio>=2.9.0

# Optional: For CLIP score evaluation
# git+https://github.com/openai/CLIP.git

# Optional: For advanced visualization
# opencv-python>=4.5.0
